add NFN Forwarder test for thunks

+++ PiCN/Layers/TimeoutPreventionLayer/BasicTimeoutPreventionLayer.py
@@ -85,6 +85,9 @@ class BasicTimeoutPreventionLayer(LayerProcess):
     def data_from_lower(self, to_lower: multiprocessing.Queue, to_higher: multiprocessing.Queue, data):
         packet_id = data[0]
         packet = data[1]
+        if 'THUNK' in str(packet.name):
+            self.queue_to_higher.put(data)
+            return
         if isinstance(packet, Interest):
             self.logger.info("Reveived Interest from lower... " + str(packet.name))
             if len(packet.name.components) > 2 and packet.name.string_components[-2] == 'KEEPALIVE':
@@ -132,6 +135,9 @@ class BasicTimeoutPreventionLayer(LayerProcess):
     def data_from_higher(self, to_lower: multiprocessing.Queue, to_higher: multiprocessing.Queue, data):
         packet_id = data[0]
         packet = data[1]
+        if 'THUNK' in str(packet.name):
+            self.queue_to_lower.put(data)
+            return
         if (isinstance(packet, Content) or isinstance(packet, Nack)) and packet.name in self.running_computations:
             self.running_computations.remove(packet.name)
             self.message_dict.remove_entry(packet.name)
